[Scope change] Extract re-usable hooks (#495)
* extract hooks

* cr fixes

* use alias
