+++ src/apps/ScopeChangeRequest/Components/PersonRoleSearch/PCSPersonRoleSearch.tsx
@@ -8,7 +8,7 @@ import {
 } from '../Inputs/SearchableDropdown/applyEds';
 import { TypedSelectOption } from '../../api/Search/searchType';
 import { sort } from '../../functions/sort';
-import { useCancellationToken } from '../../hooks/cancellationToken/useCancellationToken';
+import { useCancellationToken } from '../../../../hooks/cancellationToken/useCancellationToken';
 import { usePcsSearch } from '../../hooks/Search/usePcsSearch';
 import { useFacility } from '../../../../Core/Client/Hooks';
 import { ProCoSysQueries } from '../../keys/ProCoSysQueries';
