+++ vite.config.ts
@@ -50,6 +50,7 @@ export default defineConfig({
                 './src/packages/ModelViewer'
             ),
             '@equinor/lighthouse-status-bar': path.resolve(__dirname, './src/packages/StatusBar'),
+            '@equinor/lighthouse-hooks': path.resolve(__dirname, './src/hooks'),
         },
     },
     build: {
